Breaking up code to move mgopt to the main code branch
